fix(events): refactor GitLab events
