+++ packit_service/worker/events/gitlab/commit.py
@@ -1,0 +1,8 @@
+# Copyright Contributors to the Packit project.
+# SPDX-License-Identifier: MIT
+
+from packit_service.worker.events.comment import CommitCommentEvent
+
+
+class Comment(CommitCommentEvent):
+    pass
