+++ packit_service/worker/handlers/mixin.py
@@ -21,13 +21,15 @@ from packit_service.models import (
     SRPMBuildModel,
 )
 from packit_service.utils import get_packit_commands_from_comment
-from packit_service.worker.events import PullRequestCommentPagureEvent
+from packit_service.worker.events import (
+    MergeRequestCommentGitlabEvent,
+    PullRequestCommentPagureEvent,
+)
 from packit_service.worker.events.copr import AbstractCoprBuildEvent
 from packit_service.worker.events.event import EventData
 from packit_service.worker.events.github.pr import (
     Comment as PullRequestCommentGithubEvent,
 )
-from packit_service.worker.events.gitlab import MergeRequestCommentGitlabEvent
 from packit_service.worker.events.koji import KojiBuildEvent, KojiBuildTagEvent
 from packit_service.worker.handlers.abstract import CeleryTask
 from packit_service.worker.helpers.build.copr_build import CoprBuildJobHelper
